+++ src/App.js
@@ -1,30 +1,19 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Login from "./components/Login/Login";
-import ProtectedRoute from "./components/ProtectedRoute";
-import Register from "./components/Register";
-import AdminDash from "./components/Admin/AdminDash";
-import AdminChannel from "./components/Admin/AdminChannel";
-import MemberDash from "./components/Member/MemberDash";
-import MemberChannel from "./components/Member/MemberChannel";
-import MembersFriendsList from "./components/Member/MemberFriendsList";
-import AdminsFriendsList from "./components/Admin/AdminFriendsList";
-import PrivateChannel from "./components/PrivateChannel";
+import Register from "./components/Login/Register";
+import FriendList from "./components/FriendsList/FriendList";
+import PrivateChannel from "./components/Channels/PrivateChannel";
 import OfflineFriendList from "./components/Offline/OfflineFriendList";
-import Dashboard from "./components/Dashboard"
+import Dashboard from "./components/Dashboard/Dashboard"
 function App() {
   return (
     <Router>
       <Routes>
         <Route path="/" element={<Login />} />
         <Route path="/register" element={<Register />} />
-        <Route path="/Admin" element={<ProtectedRoute><AdminDash /></ProtectedRoute>}/>
-        <Route path="/Member" element={<ProtectedRoute><MemberDash /></ProtectedRoute>}/>
-        <Route path="/Afriends" element={<AdminsFriendsList />} />
-        <Route path="/channelA/:id" element={<AdminChannel />} />
+        <Route path="/friends" element={<FriendList />} />
         <Route path="/privchannel/:id" element={<PrivateChannel />} />
-        <Route path="/Mfriends" element={<MembersFriendsList />} />
-        <Route path="/channelM/:id" element={<MemberChannel />} />
         <Route path="/OfflineFriendList" element={<OfflineFriendList />} />
         <Route path="/Dashboard" element={<Dashboard />} />
       </Routes>
