+++ src/components/Channels/NewChannelPrompt.js
